+++ src/nr-utils.c
@@ -93,6 +93,75 @@ int *as_int32_vec(SEXP vec_, int N, bool *do_free) {
 }
 
 
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// Convert an R INTEGER or REAL to a C  "double *"
+// 
+// @param vec_ R SEXP
+// @param N Expected length
+// @param do_Free was new memory allocat in this function (the caller should free())
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+double *as_double_vec(SEXP vec_, int N, bool *do_free) {
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Must be REALSXP or INTSXP
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (!isInteger(vec_) && !isReal(vec_)) {
+    error("as_double_vec(): Expecting numeric but got %s\n", type2char((SEXPTYPE)TYPEOF(vec_)));
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Must be length 1 or N
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (length(vec_) != 1 && length(vec_) != N) {
+    error("as_double_vec(): Input vector must be length 1 or N, not %i", length(vec_));
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Do we have an integer vector of the correct length?
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (isReal(vec_) && length(vec_) == N) {
+    return REAL(vec_);
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // If we got here:
+  //   - need to allocate memory for an integer vector
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  *do_free = 1;
+  double *dbl_vec = malloc((size_t)N * sizeof(double));
+  if (dbl_vec == NULL) {
+    error("as_double_vec(): out of memory");
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // If given INTSXP
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (isReal(vec_)) {
+    // Can only be an INTSXP of length=1
+    double value = asReal(vec_);
+    for (int i = 0; i < N; i++) {
+      dbl_vec[i] = value;
+    }
+  } else {
+    // INTSXP
+    int *int_vec = INTEGER(vec_);
+    if (length(vec_) == N) {
+      for (int i = 0; i < N; i++) {
+        dbl_vec[i] = (double)(int_vec[i]);
+      }
+    } else {
+      // length == 1
+      double value = (double)int_vec[0];
+      for (int i = 0; i < N; i++) {
+        dbl_vec[i] = value;
+      }
+    }
+  }
+  
+  return dbl_vec;
+}
+
+
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Calculate group N
 //