+++ src/nr-utils.h
@@ -3,4 +3,5 @@
 
 void assert_nativeraster(SEXP nr_) ;
 int *as_int32_vec(SEXP vec_, int N, bool *do_free);
+double *as_double_vec(SEXP vec_, int N, bool *do_free);
 int calc_max_length(int count, ...);